pyproject.toml
[project]
name = "nonebot-plugin-dialectlist"
version = "2.6.2"
description = "看看你群群友有多能说"
authors = [{ name = "Chen_Xu233", email = "woyerpa@outlook.com" }]
dependencies = [
    "nonebot-plugin-chatrecorder>=0.7.0",
    "nonebot-plugin-orm[default]",
    "nonebot-plugin-apscheduler>=0.4.0",
    "nonebot-plugin-alconna>=0.50.2",
    "nonebot-plugin-cesaa>=0.4.0",
    "nonebot-plugin-userinfo>=0.2.6",
    "nonebot-plugin-htmlrender>=0.3.3",
    "nonebot2>=2.3.2",
    "pillow>=10.4.0",
    "nonebot-plugin-uninfo>=0.1.1",
]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "MIT" }


[project.optional-dependencies]
dev = [
    "ruff>=0.5.5",
    "setuptools>=71.1.0",
    "twine>=5.1.0",
    "nb-cli>=0.7.6",
    "py-spy>=0.3.14",
]
Test = ["nonebot-adapter-onebot>=2.4.4"]

[tool.pdm]
distribution = true

[tool.ruff]
line-length = 80

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"

[tool.ruff.lint]
ignore = ["E402"]
